refactor(storage): move reclamation into the systems layer (#781)
The full suite caught an architecture-boundary violation the narrower selection
missed: the manage_workspace_reclamation handler lives in brain.systems and
imported brain.jobs.pipelines.workspace_gc, so the dependency pointed the wrong
way. brain.jobs may import brain.systems, not the reverse.

Move the reclamation service to brain/systems/workspace_reclamation.py, beside
the workspace inventory it uses, and leave workspace_gc.py as a 36-line
scheduled entry point that builds a unit of work, calls the service and prints
the result. Both the job and the tool now depend on the systems layer.

The architecture allowlist is deliberately untouched — it is stale debt to be
retired, not a place to record a violation introduced the same day.

Behaviour is unchanged: every safety rule, disposition and policy read moved
intact.
